# static/openapi.yaml
openapi: 3.0.3
info:
  title: MLAIC Competition Manager API
  description: |
    REST API of the competition management system (competitors, starts, disciplines,
    results, rankings, teams and relay management).

    **Optimistic locking:** competitors, results and teams carry a `version` that is
    incremented on every save. Send the version you loaded back on update/delete; if
    someone else saved in the meantime the request is rejected with `409` and the
    response contains the `current` record.
  version: "1.0"
servers:
  - url: /api
    description: This server
  - url: http://localhost:5000/api
    description: Local development
...
paths:
...
  /ranking:
    get:
      tags: [Rankings]
      summary: Rankings of all disciplines that have results
      operationId: getAllRankings
      responses:
        "200":
          description: Rankings keyed by discipline id
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { $ref: "#/components/schemas/DisciplineRanking" }
        "500": { $ref: "#/components/responses/ServerError" }
  /ranking/{disciplineId}:
    parameters:
      - $ref: "#/components/parameters/DisciplineId"
    get:
      tags: [Rankings]
      summary: Ranking of one discipline
      description: For team disciplines this returns the team ranking (`kind` = `team`).
      operationId: getRanking
      responses:
        "200":
          description: The ranking
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/DisciplineRanking"
                  - $ref: "#/components/schemas/TeamRanking"
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
  /ranking/best:
    get:
      tags: [Rankings]
      summary: One-result rankings of all disciplines that have results
      description: Each competitor is ranked on their best result only (score, then its 10s ... 1s, then its tie-break, lower wins). Team disciplines return the team ranking.
      operationId: getAllBestResultRankings
      responses:
        "200":
          description: Rankings keyed by discipline id
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  oneOf:
                    - $ref: "#/components/schemas/BestResultDisciplineRanking"
                    - $ref: "#/components/schemas/TeamRanking"
        "500": { $ref: "#/components/responses/ServerError" }
  /ranking/best/{disciplineId}:
    parameters:
      - $ref: "#/components/parameters/DisciplineId"
    get:
      tags: [Rankings]
      summary: One-result ranking of one discipline
      description: For team disciplines this returns the team ranking (`kind` = `team`).
      operationId: getBestResultRanking
      responses:
        "200":
          description: The ranking
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/BestResultDisciplineRanking"
                  - $ref: "#/components/schemas/TeamRanking"
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }

  # ---------------------------------------------------------------------- Teams
...
  /teams/ranking/{disciplineId}:
    parameters:
      - $ref: "#/components/parameters/DisciplineId"
    get:
      tags: [Teams]
      summary: Ranked teams of a team discipline
      operationId: getTeamRanking
      responses:
        "200":
          description: The team ranking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamRanking" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
...
components:
  parameters:
...
    DisciplineId:
      name: disciplineId
      in: path
      required: true
      schema: { type: integer }
...
  responses:
...
    NotFound:
      description: Record not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    RankingDisciplineInfo:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        type: { type: string }
        unit: { type: string, example: points }
    Ranking:
      type: object
      properties:
        rank: { type: integer }
        competitor:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
            team: { type: string }
        result_totals:
          type: array
          items: { type: number }
        total_sum: { type: number }
        freq_counts:
          type: object
          description: Number of shots per ring value
          additionalProperties: { type: integer }
        override_value: { type: number, nullable: true }
        notes: { type: string }
    DisciplineRanking:
      type: object
      properties:
        kind: { type: string, enum: [individual] }
        discipline: { $ref: "#/components/schemas/RankingDisciplineInfo" }
        rankings:
          type: array
          items: { $ref: "#/components/schemas/Ranking" }
    BestResultRanking:
      type: object
      properties:
        rank: { type: integer, description: Identical results share the rank }
        competitor:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
            club: { type: string }
            country: { type: string }
        start_id: { type: string, description: Start of the best result }
        result_id: { type: integer }
        score: { type: number }
        freq_counts:
          type: object
          description: Number of shots per ring value, of the best result only
          additionalProperties: { type: integer }
        override_value: { type: number, nullable: true, description: Tie-break of the best result, lower wins }
        notes: { type: string }
    BestResultDisciplineRanking:
      type: object
      properties:
        kind: { type: string, enum: [individual] }
        discipline: { $ref: "#/components/schemas/RankingDisciplineInfo" }
        rankings:
          type: array
          items: { $ref: "#/components/schemas/BestResultRanking" }
    TeamDisciplineInfo:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        type: { type: string }
        level: { type: string }
        based_on: { type: string }
        team_size: { type: integer }
        active: { type: boolean }
        unit: { type: string, example: points }
        eligible_disciplines:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }
...
    TeamMember:
      type: object
      properties:
        start_id: { type: string }
        start_number: { type: integer }
        discipline_id: { type: integer }
        discipline_name: { type: string }
        competitor_id: { type: integer }
        name: { type: string }
        club: { type: string }
        country: { type: string }
        has_result: { type: boolean }
        score: { type: number, nullable: true }
        missing: { type: boolean, description: The start no longer exists }
    Team:
      type: object
      properties:
        id: { type: integer }
        discipline_id: { type: integer }
        discipline_name: { type: string }
        team_size: { type: integer }
        name: { type: string }
        default_name: { type: string }
        members:
          type: array
          items: { $ref: "#/components/schemas/TeamMember" }
        tie_break: { type: number, nullable: true }
        notes: { type: string }
        created_at: { type: string }
        version: { type: integer }
...
    TeamRanking:
      type: object
      properties:
        kind: { type: string, enum: [team] }
        discipline: { $ref: "#/components/schemas/TeamDisciplineInfo" }
        rankings:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/Team"
              - type: object
                properties:
                  rank: { type: integer }
                  total: { type: number }
                  freq_counts:
                    type: object
                    additionalProperties: { type: integer }
                  complete: { type: boolean, description: false if a member has no result }
...